requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
paddleocr==2.7.3
paddlepaddle==2.6.2
python-multipart==0.0.9
pillow==10.2.0
numpy==1.24.4
